backend/requirements.txt
Flask
Flask-SocketIO
python-dotenv
langchain
